First upload to repo. - Read desc for more info
This is the first upload to the repo. Some work has been done so this is
not the complete beginning but it's close enough. We have
skilaverkefni1, which is not finished at this moment and then some setup
files for a database, some tables and some data for those tables to be
used in verkefni2 and onwards.